+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   and `percentage` arguments
   * `count` has been updated to raise an error if both `by` and `over` are
   defined in the same method call
+  * `delta` has been updated to no longer accept any method arguments
 
 ## 1.5.1 (2018-06-21)